# Experiment with performance of LengthSquared