LengthSquared is one of our bottlenecks in performance. I tried to re-implement it in x86_64 assembly and benchmarked several implementations (you can find those in vec3/lsq.s). Turns out the overhead of putting all function parameter on the stack, calling our ASM pulling all the parameters off the stack, then putting the result back on the stack is a nightmare for performance. It blows away any possible performance increase. The pure Go implementation ends up being nicely inlined, and even though the ASM produced isn't the best, since it doesnt incur the cost of all those moves to/from the stack its way faster.

```
goos: linux
goarch: amd64
pkg: github.com/vfrazao-ns1/raytracing1weekend/vec3
BenchmarkLenSqASM1-24        	782524261	         1.53 ns/op	       0 B/op	       0 allocs/op
BenchmarkLenSqASM2-24        	719266898	         1.60 ns/op	       0 B/op	       0 allocs/op
BenchmarkLenSqASM3-24        	623745478	         1.89 ns/op	       0 B/op	       0 allocs/op
BenchmarkLenSqASM4-24        	688862947	         1.68 ns/op	       0 B/op	       0 allocs/op
BenchmarkLenSqASM5-24        	738215010	         1.63 ns/op	       0 B/op	       0 allocs/op
BenchmarkLenSqGo1-24         	1000000000	         0.448 ns/op	       0 B/op	       0 allocs/op
```